Convert C modules to C++ and update config
Rename several C sources to C++ (.c -> .cpp) and modernize related code; rewrite maintenance module to use SD + ArduinoJson (JSON persistence) and replace maintenance.c with maintenance.cpp. Update display code to safer spi_transaction initialization, add bounds/safety fixes and style improvements, and change pin macros to GPIO_NUM_* types. Introduce ui_dataset_populate_hist and increase history capacity in ui_data.h. Change led API to use gpio_num_t. Update platformio.ini to add ArduinoJson and SD libraries and fix upload_speed. Bump sdkconfig to ESP-IDF 5.5.3 and apply various config flag adjustments. Expose C-compatible APIs where needed (extern "C") to preserve ABI.

+++ src/components/display/ui/ui.h
@@ -95,7 +95,7 @@ static const ui_region_t BACK_REGION = {0, 0, 70, UI_BAR_H};
 #define MAINT_ROW_H 24
 #define MAINT_ROW_X 2
 #define MAINT_ROW_W (MAINT_LIST_W - 4)
-#define MAINT_ROW_Y(i) (UI_CONT_Y + 2 + (i) * (MAINT_ROW_H + 2))
+#define MAINT_ROW_Y(i) ((uint16_t)(UI_CONT_Y + 2 + (i) * (MAINT_ROW_H + 2)))
 static const ui_region_t MAINT_REG_BTN = {140, 186, 176, 24};
 
 // Configurações > Veículo — linhas e área de edição
